+++ src/inner.cpp
@@ -138,6 +138,13 @@ struct focei_options {
   double *thetaGrad = NULL;
   double *initPar = NULL;
   double *scaleC = NULL;
+  // Adaptive Gaussian Quadrature
+  //
+  double *aqx;
+  double *aqw;
+  double aqLow;
+  double aqHi;
+  bool aqfirst = false;
   double scaleC0;
   int *xPar = NULL;
   NumericVector lowerIn;
@@ -306,6 +313,9 @@ struct focei_options {
 };
 
 focei_options op_focei;
+
+int _aqn = 0;
+int _nagq = 0;
 
 struct focei_ind {
   int nInnerF;
@@ -1289,6 +1299,141 @@ arma::vec getGradForOptimHess(arma::vec &t, int id) {
 
 bool _finalObfCalc = false;
 
+bool calcEtaHessian(double *eta, int likId, int id,
+                    focei_ind *fInd,
+                    rx_solving_options_ind *ind,
+                    mat &H, mat &H0) {
+  H.zeros();
+  int k, l;
+  mat tmp;
+  // This is actually -H
+  if (op_focei.needOptimHess) {
+    arma::vec gr0(op_focei.neta);
+    std::copy(&fInd->lp[0], &fInd->lp[0] + op_focei.neta, &gr0[0]);
+
+    arma::vec grPH(op_focei.neta);
+    arma::vec grMH(op_focei.neta);
+
+    double h = 0;
+
+    for (k = op_focei.neta; k--;) {
+      h = fInd->etahh[k];
+      if (op_focei.optimHessType == 3 && h <= 0) {
+        arma::vec t(eta, op_focei.neta);
+        fInd->etahh[k] = shi21Forward(getGradForOptimHess, t, h,
+                                      gr0, grPH, id, k,
+                                      op_focei.hessEpsInner, //double ef = 7e-7,
+                                      1.5,  //double rl = 1.5,
+                                      6.0,  //double ru = 6.0);;
+                                      op_focei.shi21maxInner);  //maxiter=15
+        H.col(k) = grPH;
+        continue;
+      }
+      if (op_focei.optimHessType == 1 && h <= 0) {
+        // Central
+        arma::vec t(eta, op_focei.neta);
+        fInd->etahh[k] = shi21Central(getGradForOptimHess, t, h,
+                                      gr0, grPH, id, k,
+                                      op_focei.hessEpsInner, // ef,
+                                      1.5,//double rl = 1.5,
+                                      4.5,//double ru = 4.5,
+                                      3.0,//double nu = 8.0);
+                                      op_focei.shi21maxInner); // maxiter
+        H.col(k) = grPH;
+        continue;
+      }
+      // x + h
+      eta[k] += h;
+      lpInner(eta, &grPH[0], id);
+      bool forwardFinite =  grPH.is_finite();
+      if (op_focei.optimHessType == 3 && forwardFinite) { // forward
+        H.col(k) = (grPH-gr0)/h;
+        eta[k] -= h;
+        continue;
+      }
+
+      // x - h
+      eta[k] -= 2*h;
+      lpInner(eta, &grMH[0], id);
+      bool backwardFinite = grMH.is_finite();
+      if (op_focei.optimHessType == 1 &&
+          forwardFinite && backwardFinite) {
+        // central
+        eta[k] += h;
+        H.col(k) = (grPH-grMH)/(2.0*h);
+        continue;
+      }
+      if (forwardFinite && !backwardFinite) {
+        // forward difference
+        H.col(k) = (grPH-gr0)/h;
+        eta[k] += h;
+        continue;
+      }
+      if (!forwardFinite && backwardFinite) {
+        // backward difference
+        H.col(k) = (gr0-grMH)/h;
+        eta[k] += h;
+        continue;
+      }
+    }
+    // symmetrize
+    H = 0.5*(H + H.t());
+    // Note that since the gradient includes omegaInv*etam,
+    // op_focei.omegaInv(k, l) shouldn't be added.
+  } else if (op_focei.interaction) {
+    arma::mat a(fInd->a, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), op_focei.neta, false, true);
+    arma::mat B(fInd->B, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), 1, false, true);
+    arma::mat c(fInd->c, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), op_focei.neta, false, true);
+    for (k = op_focei.neta; k--;){
+      for (l = k+1; l--;){
+        // tmp = fInd->a.col(l) %  fInd->B % fInd->a.col(k);
+        H(k, l) = 0.5*sum(a.col(l) % B % a.col(k) +
+                          c.col(l) % c.col(k)) +
+          op_focei.omegaInv(k, l);
+        if (!R_finite(H(k, l))) {
+          return false;
+        }
+        H(l, k) = H(k, l);
+      }
+    }
+  } else {
+    arma::mat a(fInd->a, fInd->nObs, op_focei.neta, false, true);
+    // std::copy(&fInd->a[0], &fInd->a[0]+a.size(), a.begin());
+    arma::mat B(fInd->B, fInd->nObs, 1, false, true);
+    // std::copy(&fInd->B[0], &fInd->B[0]+B.size(), B.begin());
+    for (k = op_focei.neta; k--;){
+      for (l = k+1; l--;) {
+        // tmp = a.col(l) %  B % a.col(k);
+        H(k, l) = 0.5*sum(a.col(l) % B % a.col(k)) +
+          op_focei.omegaInv(k, l);
+        if (!R_finite(H(k, l))) {
+            return false;
+        }
+        H(l, k) = H(k, l);
+      }
+    }
+  }
+  k=0;
+  if (!H.is_sympd()) {
+    arma::mat H2;
+    if (nmNearPD(H2, H)) {
+      H=H2;
+    }
+  }
+  if (fInd->doChol) {
+    arma::mat Hout, Hin = H;
+    bool success = chol_sym(Hout, Hin);
+    if (!success) {
+      return false;
+    }
+    H0 = Hout;
+  } else {
+    H0=cholSE__(H, op_focei.cholSEtol);
+  }
+  H = H0.t() * H0;
+  return true;
+}
+
 double LikInner2(double *eta, int likId, int id) {
   focei_ind *fInd = &(inds_focei[id]);
   double lik=0;
@@ -1299,7 +1444,7 @@ double LikInner2(double *eta, int likId, int id) {
     lik = fInd->llik;
   } else {
     // print(wrap(op_focei.logDetOmegaInv5));
-    lik = -likInner0(eta, id) + op_focei.logDetOmegaInv5;
+    lik = -likInner0(eta, id);
     // print(wrap(lik));
     rx = getRxSolve_();
     rx_solving_options_ind *ind = getSolvingOptionsInd(rx, id);
@@ -1312,139 +1457,93 @@ double LikInner2(double *eta, int likId, int id) {
     // Calculate lik first to calculate components for Hessian
     // Hessian
     mat H(op_focei.neta, op_focei.neta);
-    H.zeros();
-    int k, l;
-    mat tmp;
-    // This is actually -H
-    if (op_focei.needOptimHess) {
-      arma::vec gr0(op_focei.neta);
-      std::copy(&fInd->lp[0], &fInd->lp[0] + op_focei.neta, &gr0[0]);
+    mat H0(op_focei.neta, op_focei.neta);
 
-      arma::vec grPH(op_focei.neta);
-      arma::vec grMH(op_focei.neta);
-
-      double h = 0;
-
-      for (k = op_focei.neta; k--;) {
-        h = fInd->etahh[k];
-        if (op_focei.optimHessType == 3 && h <= 0) {
-          arma::vec t(eta, op_focei.neta);
-          fInd->etahh[k] = shi21Forward(getGradForOptimHess, t, h,
-                                        gr0, grPH, id, k,
-                                        op_focei.hessEpsInner, //double ef = 7e-7,
-                                        1.5,  //double rl = 1.5,
-                                        6.0,  //double ru = 6.0);;
-                                        op_focei.shi21maxInner);  //maxiter=15
-          H.col(k) = grPH;
-          continue;
-        }
-        if (op_focei.optimHessType == 1 && h <= 0) {
-          // Central
-          arma::vec t(eta, op_focei.neta);
-          fInd->etahh[k] = shi21Central(getGradForOptimHess, t, h,
-                                        gr0, grPH, id, k,
-                                        op_focei.hessEpsInner, // ef,
-                                        1.5,//double rl = 1.5,
-                                        4.5,//double ru = 4.5,
-                                        3.0,//double nu = 8.0);
-                                        op_focei.shi21maxInner); // maxiter
-          H.col(k) = grPH;
-          continue;
-        }
-        // x + h
-        eta[k] += h;
-        lpInner(eta, &grPH[0], id);
-        bool forwardFinite =  grPH.is_finite();
-        if (op_focei.optimHessType == 3 && forwardFinite) { // forward
-          H.col(k) = (grPH-gr0)/h;
-          eta[k] -= h;
-          continue;
-        }
-
-        // x - h
-        eta[k] -= 2*h;
-        lpInner(eta, &grMH[0], id);
-        bool backwardFinite = grMH.is_finite();
-        if (op_focei.optimHessType == 1 &&
-            forwardFinite && backwardFinite) {
-          // central
-          eta[k] += h;
-          H.col(k) = (grPH-grMH)/(2.0*h);
-          continue;
-        }
-        if (forwardFinite && !backwardFinite) {
-          // forward difference
-          H.col(k) = (grPH-gr0)/h;
-          eta[k] += h;
-          continue;
-        }
-        if (!forwardFinite && backwardFinite) {
-          // backward difference
-          H.col(k) = (gr0-grMH)/h;
-          eta[k] += h;
-          continue;
-        }
-      }
-      // symmetrize
-      H = 0.5*(H + H.t());
-      // Note that since the gradient includes omegaInv*etam,
-      // op_focei.omegaInv(k, l) shouldn't be added.
-    } else if (op_focei.interaction) {
-      arma::mat a(fInd->a, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), op_focei.neta, false, true);
-      arma::mat B(fInd->B, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), 1, false, true);
-      arma::mat c(fInd->c, getIndNallTimes(ind) - getIndNdoses(ind) - getIndNevid2(ind), op_focei.neta, false, true);
-      for (k = op_focei.neta; k--;){
-        for (l = k+1; l--;){
-          // tmp = fInd->a.col(l) %  fInd->B % fInd->a.col(k);
-          H(k, l) = 0.5*sum(a.col(l) % B % a.col(k) +
-                            c.col(l) % c.col(k)) +
-            op_focei.omegaInv(k, l);
-          if (!R_finite(H(k, l))) return NA_REAL;
-          H(l, k) = H(k, l);
-        }
-      }
-    } else {
-      arma::mat a(fInd->a, fInd->nObs, op_focei.neta, false, true);
-      // std::copy(&fInd->a[0], &fInd->a[0]+a.size(), a.begin());
-      arma::mat B(fInd->B, fInd->nObs, 1, false, true);
-      // std::copy(&fInd->B[0], &fInd->B[0]+B.size(), B.begin());
-      for (k = op_focei.neta; k--;){
-        for (l = k+1; l--;) {
-          // tmp = a.col(l) %  B % a.col(k);
-          H(k, l) = 0.5*sum(a.col(l) % B % a.col(k)) +
-            op_focei.omegaInv(k, l);
-          if (!R_finite(H(k, l))) return NA_REAL;
-          H(l, k) = H(k, l);
-        }
-      }
+    if (!calcEtaHessian(eta, likId, id, fInd, ind, H, H0)) {
+      return NA_REAL;
     }
-    arma::mat H0(op_focei.neta, op_focei.neta);
-    k=0;
-    if (!H.is_sympd()) {
-      arma::mat H2;
-      if (nmNearPD(H2, H)) {
-        H=H2;
-      }
-    }
-    if (fInd->doChol) {
-      arma::mat Hout, Hin = H;
-      bool success = chol_sym(Hout, Hin);
-      if (!success) {
-        return NA_REAL;
-      }
-      H0 = Hout;
-    } else {
-      H0=cholSE__(H, op_focei.cholSEtol);
-    }
-    H = H0.t() * H0;
     if (_finalObfCalc) {
       std::copy(H.begin(), H.end(),
                 op_focei.gH + id*op_focei.neta*op_focei.neta);
     }
-
     // - sum(log(H.diag()));
+    double logH0diag = 0.0;
     for (unsigned int j = H0.n_rows; j--;){
-      lik -= _safe_log(H0(j,j));
+      logH0diag -= _safe_log(H0(j,j));
+    }
+    if (_aqn == 0) {
+      lik += logH0diag + op_focei.logDetOmegaInv5;
+    } else {
+      // This is where the log likelihood can be adapted for the
+      // Gaussian Hermite Quadrature
+      // https://en.wikipedia.org/wiki/Gauss%E2%80%93Hermite_quadrature
+
+      // At the mode of the distribution the eta is given by the EBE
+      // estimates (determined above).
+      //
+      // The standard error of the EBE estimates is given by the inverse
+      // of the cholesky decomposition of the hessian
+      //
+      // Hence this is used in the expansion of the likelihood for the
+      // adaptive Gaussian Hermite Quadrature.
+
+      // The standard error is not needed for the Laplace approximation
+
+      // Get x and w from the AQ
+      arma::mat aqx(op_focei.aqx, _aqn, op_focei.neta, false, true);
+      arma::mat aqw(op_focei.aqw, _aqn, op_focei.neta, false, true);
+
+      // llik = -likInner0(eta, id);
+      int curi = 0;
+      double slik = 0;
+      if (op_focei.aqfirst) {
+        // Already calculated:
+        // lik = -likInner0(eta, id);
+        arma::vec w = aqw.row(curi).t();
+        curi=1;
+        lik += sum(log(w)); // x % x  = x^2; here x=0
+        // can be factored out
+        //lik += op_focei.logDetOmegaInv5;
+        lik = max2(lik, op_focei.aqLow);
+        lik = min2(lik, op_focei.aqHi);
+        slik = exp(lik);
+      }
+      arma::mat Ginv_5(op_focei.neta, op_focei.neta);
+      if (_aqn > 1) {
+        // The Ginv_5 only needs to be calculated for the
+        // non-Laplace case
+        bool success;
+        try {
+          success = inv(Ginv_5, trimatu(H0));
+          if (!success) {
+            success = inv(Ginv_5, H0);
+            if (!success) {
+              return NA_REAL;
+            }
+          }
+        } catch (...) {
+          success = inv(Ginv_5, H0);
+          if (!success) {
+            return NA_REAL;
+          }
+        }
+        arma::vec etahat(eta, op_focei.neta);
+        for (; curi < _aqn; curi++) {
+          // Get the x and w for the current iteration
+          arma::vec x = aqx.row(curi).t();
+          arma::vec w = aqw.row(curi).t();
+          arma::vec etaCur = etahat +  Ginv_5 * x;
+          lik  = -likInner0(etaCur.memptr(), id);
+          lik += sum(log(w) +  0.5 * x % x); // x % x  = x^2
+          // Can be factored out
+          //lik += op_focei.logDetOmegaInv5;
+          lik = max2(lik, op_focei.aqLow);
+          lik = min2(lik, op_focei.aqHi);
+          slik += exp(lik);
+        }
+      }
+      //lik = 0.5*op_focei.neta * M_LN2 + det_Ginv_5 + log(slik);
+      lik = log(slik) + logH0diag + op_focei.logDetOmegaInv5;
     }
   }
   lik += fInd->tbsLik;
@@ -2986,10 +3085,13 @@ static inline void foceiSetupTrans_(CharacterVector pars){
   op_focei.etaFD       = op_focei.fixedTrans + op_focei.ntheta + op_focei.omegan; // [neta]
 
   if (op_focei.fullTheta != NULL) R_Free(op_focei.fullTheta);
-  op_focei.fullTheta   = R_Calloc(4*(op_focei.ntheta+op_focei.omegan), double); // [ntheta+omegan]
+  op_focei.fullTheta   = R_Calloc(4*(op_focei.ntheta+op_focei.omegan) +
+                                  2*(_aqn*op_focei.neta), double); // [ntheta+omegan]
   op_focei.theta       = op_focei.fullTheta+op_focei.ntheta+op_focei.omegan; // [ntheta + omegan]
   op_focei.initPar     = op_focei.theta+op_focei.ntheta+op_focei.omegan; // [ntheta + omegan]
   op_focei.scaleC      = op_focei.initPar+op_focei.ntheta+op_focei.omegan; // [ntheta + omegan]
+  op_focei.aqx         = op_focei.scaleC + op_focei.ntheta+op_focei.omegan; // [aqn*neta]
+  op_focei.aqw         = op_focei.aqx + _aqn*op_focei.neta; // [aqn*neta]
 
   int neta = 0;
   unsigned int ntheta = 0;
@@ -5273,6 +5375,18 @@ NumericMatrix sqrtm(NumericMatrix m){
   return wrap(re);
 }
 
+
+void setupAq0_(Environment e) {
+  if (e.exists("aqn")) {
+    _aqn = as<int>(e["aqn"]);
+    _nagq = as<int>(e["nAGQ"]);
+  } else {
+    _aqn = 0;
+    _nagq = 0;
+  }
+}
+
+
 //[[Rcpp::export]]
 NumericMatrix foceiCalcCov(Environment e){
   std::string boundStr = "";
@@ -5357,6 +5471,7 @@ NumericMatrix foceiCalcCov(Environment e){
         op_focei.lastOfv = op_focei.lastOfv * op_focei.initObjective / op_focei.scaleObjectiveTo;
       }
       // foceiSetupTheta_(op_focei.mvi, fullT2, skipCov, op_focei.scaleTo, false);
+      setupAq0_(e);
       foceiSetupTheta_(op_focei.mvi, fullT2, skipCov, 0, false);
       op_focei.scaleType=10;
       if (op_focei.covMethod && !boundary) {
@@ -6373,6 +6488,14 @@ void foceiFinalizeTables(Environment e){
     std::string ofvType = as<std::string>(e["ofvType"]);
     objDf.attr("row.names") = ofvType;
     e["ofvType"]= ofvType;
+  } else if (_nagq > 0)  {
+    if (_nagq == 1) {
+      objDf.attr("row.names") = CharacterVector::create("Laplace");
+    } else {
+      objDf.attr("row.names") = CharacterVector::create("AGQ" + std::to_string(_nagq));
+    }
+    addLlikObs(e);
+    e["ofvType"] = "agq";
   } else {
     if (op_focei.needOptimHess) {
       objDf.attr("row.names") = CharacterVector::create("lFOCEi");
@@ -6385,7 +6508,9 @@ void foceiFinalizeTables(Environment e){
   objDf.attr("class") = "data.frame";
   e["objDf"]=objDf;
   if (!e.exists("method")){
-    if (op_focei.neta == 0){
+    if (_aqn > 0) {
+      e["method"] ="AGQ";
+    } else if (op_focei.neta == 0){
       e["method"] = "Population Only";
     } else if (op_focei.fo == 1){
       e["method"] = "FO";
@@ -6410,10 +6535,36 @@ void foceiFinalizeTables(Environment e){
       e["extra"] = "";
     }
     List ctl = e["control"];
-    e["extra"] = as<std::string>(e["extra"]) + " (outer: " + as<std::string>(ctl["outerOptTxt"]) +")";
+    if (_aqn == 0) {
+      e["extra"] = as<std::string>(e["extra"]) +
+        " (outer: " + as<std::string>(ctl["outerOptTxt"]) +
+        ")";
+    } else if (_aqn == 1) {
+      e["extra"] = as<std::string>(e["extra"]) +
+        " (outer: " + as<std::string>(ctl["outerOptTxt"]) +
+        "; Laplace)";
+    } else {
+      e["extra"] = as<std::string>(e["extra"]) +
+        " (outer: " + as<std::string>(ctl["outerOptTxt"]) +
+        "; nAGQ=" + std::to_string(_nagq)  + ")";
+    }
   }
   // rxode2::rxSolveFree();
   e.attr("class") = "nlmixr2FitCore";
+}
+
+void setupAq1_(Environment e) {
+  if (_aqn >= 1) {
+    double *tmp = REAL(e["qx"]);
+    std::copy(tmp, tmp + _aqn*op_focei.neta, op_focei.aqx);
+    tmp = REAL(e["qw"]);
+    std::copy(tmp, tmp + _aqn*op_focei.neta, op_focei.aqw);
+    op_focei.aqfirst = as<bool>(e["qfirst"]);
+    op_focei.aqLow = as<double>(e["aqLow"]);
+    op_focei.aqHi = as<double>(e["aqHi"]);
+  } else {
+    op_focei.aqfirst = false;
+  }
 }
 
 //' Fit/Evaluate FOCEi
@@ -6433,10 +6584,11 @@ Environment foceiFitCpp_(Environment e){
   bool doPredOnly = false;
   op_focei.canDoFD = false;
   if (e.exists("nEstOmega")) {
-    op_focei.nEstOmega = e["nEstOmega"];
+    op_focei.nEstOmega = as<int>(e["nEstOmega"]);
   } else {
     op_focei.nEstOmega = 0;
   }
+  setupAq0_(e);
   if (model.containsElementNamed("inner")) {
     RObject inner;
     if (model.containsElementNamed("innerLlik")) {
@@ -6454,6 +6606,7 @@ Environment foceiFitCpp_(Environment e){
       Nullable<NumericVector> _upper = as<Nullable<NumericVector>> (e["upper"]);
       Nullable<NumericMatrix> _etaMat = as<Nullable<NumericMatrix>>(e["etaMat"]);
       Nullable<List> _control = as<Nullable<List>>(e["control"]);
+      setupAq0_(e);
       foceiSetup_(inner, _dataSav, _thetaIni, _thetaFixed, _skipCov,
                   _rxInv, _lower, _upper, _etaMat, _control);
       if (model.containsElementNamed("predNoLhs")) {
@@ -6494,6 +6647,7 @@ Environment foceiFitCpp_(Environment e){
         Nullable<NumericVector> _upper = as<Nullable<NumericVector>> (e["upper"]);
         Nullable<NumericMatrix> _etaMat = as<Nullable<NumericMatrix>>(e["etaMat"]);
         Nullable<List> _control = as<Nullable<List>>(e["control"]);
+        setupAq0_(e);
         foceiSetup_(inner, _dataSav, _thetaIni, _thetaFixed, _skipCov,
                     _rxInv, _lower, _upper, _etaMat, _control);
         doPredOnly = true;
@@ -6503,6 +6657,7 @@ Environment foceiFitCpp_(Environment e){
       }
     } else {
       doPredOnly=true;
+      setupAq0_(e);
       foceiSetupTrans_(as<CharacterVector>(e[".params"]));
       RObject _dataSav = as<RObject>(e["dataSav"]);
       NumericVector _thetaIni = as<NumericVector>(e["thetaIni"]);
@@ -6513,6 +6668,7 @@ Environment foceiFitCpp_(Environment e){
       Nullable<NumericVector> _upper = as<Nullable<NumericVector>> (e["upper"]);
       Nullable<NumericMatrix> _etaMat = as<Nullable<NumericMatrix>>(e["etaMat"]);
       Nullable<List> _control = as<Nullable<List>>(e["control"]);
+      setupAq0_(e);
       foceiSetup_(inner, _dataSav, _thetaIni, _thetaFixed, _skipCov,
                   _rxInv, _lower, _upper, _etaMat, _control);
     }
@@ -6565,6 +6721,7 @@ Environment foceiFitCpp_(Environment e){
       Nullable<NumericVector> _upper = as<Nullable<NumericVector>> (e["upper"]);
       Nullable<NumericMatrix> _etaMat = as<Nullable<NumericMatrix>>(e["etaMat"]);
       Nullable<List> _control = as<Nullable<List>>(e["control"]);
+      setupAq0_(e);
       foceiSetup_(inner, _dataSav, _thetaIni, _thetaFixed, _skipCov,
                   _rxInv, _lower, _upper, _etaMat, _control);
       doPredOnly = true;
@@ -6573,6 +6730,7 @@ Environment foceiFitCpp_(Environment e){
       stop(_("model$predOnly needs to be an rxode2 object"));
     }
   }
+  setupAq1_(e);
   if (e.exists("setupTime")){
     e["setupTime"] = as<double>(e["setupTime"])+(((double)(clock() - t0))/CLOCKS_PER_SEC);
   } else {